Optimised queue component
